Export world

See merge request 11serega22/agar-life!4

+++ object/alive/animal/behavior/testangel.go
@@ -17,8 +17,8 @@ func NewTestAngel(angel float64) animal.Behavior {
 	return &testAngel{angel: angel}
 }
 
-func (a *testAngel) Action(self animal.Animal, animals []alive.Alive, plants []alive.Alive, cycle uint64) (crd.Crd, bool) {
-	//vec := vector.GetVectorByPoint(self.GetCrd(), crd.NewCrd(self.X()+self.Vision(), self.Y()))
+func (a *testAngel) Action(self animal.Animal, animals []alive.Alive, plants []alive.Alive, cycle uint) (crd.Crd, bool) {
+	//vec := vector.GetVectorByPoint(self.GetCrd(), crd.NewCrd(self.GetX()+self.GetVision(), self.GetY()))
 	//vec.SetAngle(a.angel)
 	//a.direction.SetCrd(vec.GetPointFromVector(self.GetCrd()))
 	cx := 400.0
